trae/lost_and_found/pyproject.toml
[project]
name = "lost-and-found"
version = "0.1.0"
description = "API后端服务，用于查询宿舍电费信息"
readme = "README.md"
requires-python = ">=3.14"
dependencies = [
    "flask>=3.1.3",
    "numpy>=2.4.3",
    "pillow>=12.1.1",
    "pytesseract>=0.3.13",
    "requests>=2.32.5",
]
